mediacapture-streams/stream-api/mediastream/mediastream-currenttime.html

<!DOCTYPE html>
<html>
<head>
<title>MediaStream: test MediaStream.currentTime</title>
<link rel="author" title="Kyle Barrow" href="mailto:kyle@barrow.io">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html">
<meta name="assert" content="the current time of the media stream">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">

<style>

video {
  width: 50%;
}

</style>
</head>
<body>
<p>When prompted, <strong>please allow</strong> access to the video stream.</p>
<div id="log"></div>
<script>

var t = async_test("Test MediaStream.currentTime", {timeout:10000});

//navigator.getUserMedia = navigator.getUserMedia || navigator.webkitGetUserMedia || navigator.mozGetUserMedia

t.step(function() {
	navigator.getUserMedia(
		{video: true},
		t.step_func(function(stream) {
			assert_true(stream.currentTime >= 0, "Test MediaStream.currentTime is an interger >= 0")
			stream.currentTime = 1000
			assert_true(stream.currentTime < 1000, "Test MediaStream.currentTime is read only")
			t.done();
		}),
		t.step_func(function() {
			assert_unreached("The error callback should not have been triggered as user should allow access");
			t.done();
		 })
	);
});



</script>
</body>
</html>
